completed all smal field changes and done calculation logic

+++ ERP/src/main/java/com/erp/Dto/Request/QuotationProductRequestDto.java
@@ -5,13 +5,15 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Builder
 @AllArgsConstructor
-public class QuotationProductRequestDto {
+public class QuotationProductRequestDto  {
 
     private Long productId;
-    private Integer quantity;
+    private Long quantity;
 
 }
