+++ ERP/src/main/java/com/erp/Dto/Request/SaledOrderProductRequestDto.java
@@ -1,22 +1,23 @@
 package com.erp.Dto.Request;
 
-import jakarta.validation.constraints.DecimalMin;
-import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 import java.math.BigDecimal;
 
 @Data
-public class SaledOrderProductRequestDto {
+public class SaledOrderProductRequestDto  {
 
 
     private Long serviceOrProductId;
 
-    private BigDecimal quantity;
+    private Long quantity;
 
     private BigDecimal subtotal;
 
     private BigDecimal taxAmount;
 
     private BigDecimal totalAmount;
+
+
+
 }
